pyproject.toml
[tool.poetry]
name = "sales-late-day-tokens-service-create-order"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pika = "^1.3.2"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
sqlalchemy = "^2.0.23"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
